requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
email-validator>=1.1.3
python-dotenv>=0.19.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
PyJWT>=2.0.0

